% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_operations.R
\name{delete_cluster}
\alias{delete_cluster}
\title{Deletes a previously provisioned DAX cluster}
\usage{
delete_cluster(ClusterName)
}
\arguments{
\item{ClusterName}{[required] The name of the cluster to be deleted.}
}
\description{
Deletes a previously provisioned DAX cluster. \emph{DeleteCluster} deletes all associated nodes, node endpoints and the DAX cluster itself. When you receive a successful response from this action, DAX immediately begins deleting the cluster; you cannot cancel or revert this action.
}
\section{Accepted Parameters}{
\preformatted{delete_cluster(
  ClusterName = "string"
)
}
}

