% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_operations.R
\name{delete_parameter_group}
\alias{delete_parameter_group}
\title{Deletes the specified parameter group}
\usage{
delete_parameter_group(
  ParameterGroupName = "string"
)
}
\arguments{
\item{ParameterGroupName}{[required] The name of the parameter group to delete.}
}
\description{
Deletes the specified parameter group. You cannot delete a parameter group if it is associated with any DAX clusters.
}
\examples{

}
