% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_operations.R
\name{delete_subnet_group}
\alias{delete_subnet_group}
\title{Deletes a subnet group}
\usage{
delete_subnet_group(SubnetGroupName)
}
\arguments{
\item{SubnetGroupName}{[required] The name of the subnet group to delete.}
}
\description{
Deletes a subnet group.
}
\details{
You cannot delete a subnet group if it is associated with any DAX clusters.
}
\section{Accepted Parameters}{
\preformatted{delete_subnet_group(
  SubnetGroupName = "string"
)
}
}

