% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_operations.R
\name{describe_default_parameters}
\alias{describe_default_parameters}
\title{Returns the default system parameter information for the DAX caching software}
\usage{
describe_default_parameters(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to include in the response. If more results exist than the specified \code{MaxResults} value, a token is included in the response so that the remaining results can be retrieved.

The value for \code{MaxResults} must be between 20 and 100.}

\item{NextToken}{An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by \code{MaxResults}.}
}
\description{
Returns the default system parameter information for the DAX caching software.
}
\examples{

}
