% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_package.R
\docType{package}
\name{paws.dax-package}
\alias{paws.dax}
\alias{paws.dax-package}
\title{paws.dax: Amazon DynamoDB Accelerator (DAX)}
\description{
DAX is a managed caching service engineered for Amazon
DynamoDB. DAX dramatically speeds up database reads by caching
frequently-accessed data from DynamoDB, so applications can access
that data with sub-millisecond latency. You can create a DAX cluster
easily, using the AWS Management Console. With a few simple
modifications to your code, your application can begin taking
advantage of the DAX cluster and realize significant improvements in
read performance.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
