% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_operations.R
\name{reboot_node}
\alias{reboot_node}
\title{Reboots a single node of a DAX cluster}
\usage{
reboot_node(ClusterName, NodeId)
}
\arguments{
\item{ClusterName}{The name of the DAX cluster containing the node to be rebooted.}

\item{NodeId}{The system-assigned ID of the node to be rebooted.}
}
\description{
Reboots a single node of a DAX cluster. The reboot action takes place as soon as possible. During the reboot, the node status is set to REBOOTING.
}
\examples{

}
