% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_operations.R
\name{update_subnet_group}
\alias{update_subnet_group}
\title{Modifies an existing subnet group}
\usage{
update_subnet_group(SubnetGroupName, Description = NULL,
  SubnetIds = NULL)
}
\arguments{
\item{SubnetGroupName}{[required] The name of the subnet group.}

\item{Description}{A description of the subnet group.}

\item{SubnetIds}{A list of subnet IDs in the subnet group.}
}
\description{
Modifies an existing subnet group.
}
\section{Accepted Parameters}{
\preformatted{update_subnet_group(
  SubnetGroupName = "string",
  Description = "string",
  SubnetIds = list(
    "string"
  )
)
}
}

