% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_operations.R
\name{create_subnet_group}
\alias{create_subnet_group}
\title{Creates a new subnet group}
\usage{
create_subnet_group(SubnetGroupName, SubnetIds, Description = NULL)
}
\arguments{
\item{SubnetGroupName}{A name for the subnet group. This value is stored as a lowercase string.}

\item{SubnetIds}{A list of VPC subnet IDs for the subnet group.}

\item{Description}{A description for the subnet group}
}
\description{
Creates a new subnet group.
}
\examples{

}
