% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.dax_operations.R
\name{update_parameter_group}
\alias{update_parameter_group}
\title{Modifies the parameters of a parameter group}
\usage{
update_parameter_group(ParameterGroupName, ParameterNameValues)
}
\arguments{
\item{ParameterGroupName}{The name of the parameter group.}

\item{ParameterNameValues}{An array of name-value pairs for the parameters in the group. Each element in the array represents a single parameter.}
}
\description{
Modifies the parameters of a parameter group. You can modify up to 20 parameters in a single request by submitting a list parameter name and value pairs.
}
\examples{

}
